+++ vite.config.ts
@@ -1,7 +1,14 @@
-import { sveltekit } from '@sveltejs/kit/vite';
-import tailwindcss from '@tailwindcss/vite';
 import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import tailwindcss from '@tailwindcss/vite';
+import { resolve } from 'path';
 
 export default defineConfig({
-	plugins: [tailwindcss(), sveltekit()]
+  plugins: [react(), tailwindcss()],
+  resolve: {
+    alias: {
+      $lib: resolve(__dirname, 'src/lib'),
+    },
+  },
+  assetsInclude: ['**/*.md'],
 });